requirements.txt
streamlit>=1.50.0
openai>=2.1.0
pandas
scikit-learn
pillow>=11.3.0
python-dotenv>=1.1.1
requests>=2.32.5
pytrends>=4.9.2
